NAS-114893 / 22.02.1 / Add debian-security to builder

+++ scale_build/bootstrap/bootstrapdir.py
@@ -52,13 +52,6 @@ class BootstrapDir(CacheMixin, HashMixin):
         with open(os.path.join(apt_path, 'preferences'), 'w') as f:
             f.write(get_apt_preferences())
 
-        self.logger.debug('Adding debian-security to sources')
-        deb_security = '\n'.join([
-            'deb http://deb.debian.org/debian-security/ bullseye-security main',
-            'deb-src http://deb.debian.org/debian-security/ bullseye-security main',
-        ])
-        with open(apt_sources_path, 'a+') as f:
-            f.write(f'\n{deb_security}\n')
         run(['chroot', self.chroot_basedir, 'apt', 'update'])
 
         if self.extra_packages_to_install:
@@ -74,9 +67,11 @@ class BootstrapDir(CacheMixin, HashMixin):
         # Add additional repos
         for repo in apt_repos['additional']:
             self.logger.debug('Adding additional repo: %r', repo['url'])
-            shutil.copy(os.path.join(BUILDER_DIR, repo['key']), os.path.join(self.chroot_basedir, 'apt.key'))
-            run(['chroot', self.chroot_basedir, 'apt-key', 'add', '/apt.key'])
-            os.unlink(os.path.join(self.chroot_basedir, 'apt.key'))
+            if repo.get('key'):
+                shutil.copy(os.path.join(BUILDER_DIR, repo['key']), os.path.join(self.chroot_basedir, 'apt.key'))
+                run(['chroot', self.chroot_basedir, 'apt-key', 'add', '/apt.key'])
+                os.unlink(os.path.join(self.chroot_basedir, 'apt.key'))
+
             apt_sources.append(f'deb {repo["url"]} {repo["distribution"]} {repo["component"]}')
 
         with open(apt_sources_path, 'w') as f:
